+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- New prop to `UserAddress` showIfMasked to remove entire string if asterics are present
+
 ## [3.158.0] - 2022-03-23
 
 ### Added